added ability map permissions to grants and created centralized service to update ACLs. running into db issues when spring acl is trying to create records

+++ src/main/java/com/alexbiehl/demo/service/AclService.java
@@ -1,23 +1,39 @@
 package com.alexbiehl.demo.service;
 
-import com.alexbiehl.demo.model.DBItemBase;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.acls.domain.BasePermission;
 import org.springframework.security.acls.domain.GrantedAuthoritySid;
 import org.springframework.security.acls.domain.ObjectIdentityImpl;
+import org.springframework.security.acls.domain.PrincipalSid;
 import org.springframework.security.acls.model.*;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.Serializable;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Service
+@Transactional
 public class AclService {
 
     @Autowired
     private MutableAclService aclService;
 
+
+    /**
+     * Return list of permissions in order that we like
+     * @return - list of permissions in order
+     */
+    public Permission[] getPermissions() {
+        return new Permission[]{
+                BasePermission.CREATE,
+                BasePermission.READ,
+                BasePermission.WRITE,
+                BasePermission.DELETE,
+                BasePermission.ADMINISTRATION
+        };
+    }
 
     private AuditableAcl getOrCreateAcl(String type, Serializable id) {
         AuditableAcl acl = getAcl(type, id);
@@ -39,20 +55,27 @@ public class AclService {
         }
     }
 
-    public AuditableAcl grantPermissionsToSid(String type, Serializable id, String sidName,
-                                              Permission[] permissions) {
-        return grantPermissions(type, id, new GrantedAuthoritySid(sidName), permissions);
+    public void grantPermissionsToPrincipal(String type, Serializable id, String name,
+                                            Permission[] permissions, Boolean[] grants) {
+        grantPermissions(type, id, new PrincipalSid(name), permissions, grants);
     }
 
-    private AuditableAcl grantPermissions(String type, Serializable id, Sid sid, Permission[] permissions) {
+    public void grantPermissionsToSid(String type, Serializable id, String sidName,
+                                      Permission[] permissions, Boolean[] grants) {
+        grantPermissions(type, id, new GrantedAuthoritySid(sidName), permissions, grants);
+    }
+
+    private AuditableAcl grantPermissions(String type, Serializable id, Sid sid, Permission[] permissions, Boolean[] grants) {
         AuditableAcl acl = getOrCreateAcl(type, id);
         Set<Integer> indices = new HashSet<>();
 
+        int grantIndex = 0;
         for (Permission permission : permissions) {
             int index = acl.getEntries().size();
-            boolean granting = true;
+            boolean granting = grants[grantIndex];
             acl.insertAce(index, permission, sid, granting);
             indices.add(index);
+            ++grantIndex;
         }
 
         for (Integer index : indices) {
@@ -70,8 +93,8 @@ public class AclService {
             boolean deletedEntry = false;
             for (Permission permission : permissions) {
                 if (entry.isGranting() &&
-                entry.getSid().equals(sid) &&
-                entry.getPermission().equals(permission)) {
+                        entry.getSid().equals(sid) &&
+                        entry.getPermission().equals(permission)) {
                     acl.deleteAce(index);
                     deletedEntry = true;
                 }
